New arrows logic + new progression

+++ Assets/Game/Code/Global/Settings/Core/CoreSpeedProgression.cs
@@ -5,7 +5,7 @@ namespace Global.Settings.Core
     [Serializable]
     public class CoreSpeedProgression
     {
-        public int scores;
+        public ulong scores;
         public float speed;
     }
 }
